Adicionadas novas exceptions personalizadas e criados testes de unidade

+++ src/main/java/com/lambdateam/mycar/service/ComponentService.java
@@ -6,38 +6,59 @@ import com.lambdateam.mycar.repository.ComponentRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import com.lambdateam.mycar.exception.ExpiredJwtException;;
+
 @AllArgsConstructor
 @Service
 public class ComponentService {
 
     private final ComponentRepository repository;
 
-    private ComponentModel findOrThrow(final Long id) {
-        return repository
-                .findById(id)
-                .orElseThrow(
-                        () -> new NotFoundException("The component id " + id +" was not found.")
-                );
+    private ComponentModel findOrThrow(final Long id) throws ExpiredJwtException {
+        try {
+            return repository
+                    .findById(id)
+                    .orElseThrow(
+                            () -> new NotFoundException("The component id " + id +" was not found.")
+                    );
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public Iterable<ComponentModel> findAllComponents() {
-        return repository.findAll();
+    public Iterable<ComponentModel> findAllComponents() throws ExpiredJwtException {
+        try {
+            return repository.findAll();
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public ComponentModel findComponentById(Long id) {
+    public ComponentModel findComponentById(Long id) throws ExpiredJwtException {
+        if(!repository.existsById(id)) throw new NotFoundException("The component id " + id +" was not found.");
         return findOrThrow(id);
     }
 
     public void deleteComponentById(Long id) {
+        if(!repository.existsById(id)) throw new NotFoundException("The component id " + id +" was not found.");
         repository.deleteById(id);
     }
 
-    public ComponentModel createComponent(ComponentModel component) {
-        return repository.save(component);
+    public ComponentModel createComponent(ComponentModel component) throws ExpiredJwtException {
+        try{
+            return repository.save(component);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public void updateComponent(Long id, ComponentModel component) {
-        findOrThrow(id);
-        repository.save(component);
+    public void updateComponent(Long id, ComponentModel component) throws ExpiredJwtException {
+        try{
+            if(!repository.existsById(id)) throw new NotFoundException("The component id " + id +" was not found.");
+            findOrThrow(id);
+            repository.save(component);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 }
