+++ src/main/java/com/lambdateam/mycar/service/ShoppingListService.java
@@ -1,5 +1,6 @@
 package com.lambdateam.mycar.service;
 
+import com.lambdateam.mycar.exception.ExpiredJwtException;
 import com.lambdateam.mycar.exception.NotFoundException;
 import com.lambdateam.mycar.model.ShoppingListModel;
 import com.lambdateam.mycar.repository.ShoppingListRepository;
@@ -14,36 +15,67 @@ public class ShoppingListService {
 
     private final ShoppingListRepository repository;
 
-    private ShoppingListModel findOrThrow(final Long id) {
-        return repository
-                .findById(id)
-                .orElseThrow(
-                        () -> new NotFoundException("The shopping list item " + id +" was not found.")
-                );
+    private ShoppingListModel findOrThrow(final Long id) throws ExpiredJwtException {
+        try{
+            return repository
+                    .findById(id)
+                    .orElseThrow(
+                            () -> new NotFoundException("The shopping list item " + id +" was not found.")
+                    );
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public List<ShoppingListModel> findAllShoppingListWithDetails() {
-        return repository.getShoppingListWithDetails();
+    public List<ShoppingListModel> findAllShoppingListWithDetails() throws ExpiredJwtException {
+        try{
+            return repository.getShoppingListWithDetails();
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public Iterable<ShoppingListModel> findAllShoppingListItems() {
-        return repository.findAll();
+    public Iterable<ShoppingListModel> findAllShoppingListItems() throws ExpiredJwtException {
+        try{
+            return repository.findAll();
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public ShoppingListModel findShoppingListItemById(Long id) {
-        return findOrThrow(id);
+    public ShoppingListModel findShoppingListItemById(Long id) throws ExpiredJwtException {
+        try{
+            if(!repository.existsById(id)) throw new NotFoundException("The shopping list item " + id +" was not found.");
+            return findOrThrow(id);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public void deleteShoppingListItemById(Long id) {
-        repository.deleteById(id);
+    public void deleteShoppingListItemById(Long id) throws ExpiredJwtException {
+        try{
+            if(!repository.existsById(id)) throw new NotFoundException("The shopping list item " + id +" was not found.");
+            repository.deleteById(id);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public ShoppingListModel createShoppingListItem(ShoppingListModel shoppingListItem) {
-        return repository.save(shoppingListItem);
+    public ShoppingListModel createShoppingListItem(ShoppingListModel shoppingListItem) throws ExpiredJwtException {
+        try{
+            return repository.save(shoppingListItem);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public void updateShoppingListItem(Long id, ShoppingListModel shoppingListItem) {
-        findOrThrow(id);
-        repository.save(shoppingListItem);
+    public void updateShoppingListItem(Long id, ShoppingListModel shoppingListItem) throws ExpiredJwtException {
+        try{
+            if(!repository.existsById(id)) throw new NotFoundException("The shopping list item " + id +" was not found.");
+            findOrThrow(id);
+            repository.save(shoppingListItem);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 }
