+++ src/main/java/com/lambdateam/mycar/service/WorkshopService.java
@@ -1,5 +1,6 @@
 package com.lambdateam.mycar.service;
 
+import com.lambdateam.mycar.exception.ExpiredJwtException;
 import com.lambdateam.mycar.exception.NotFoundException;
 import com.lambdateam.mycar.model.WorkshopModel;
 import com.lambdateam.mycar.repository.WorkshopRepository;
@@ -12,32 +13,59 @@ public class WorkshopService {
 
     private final WorkshopRepository repository;
 
-    private WorkshopModel findOrThrow(final Long id) {
-        return repository
-                .findById(id)
-                .orElseThrow(
-                        () -> new NotFoundException("The workshop id " + id +" was not found.")
-                );
+    private WorkshopModel findOrThrow(final Long id) throws ExpiredJwtException {
+        try{
+            return repository
+                    .findById(id)
+                    .orElseThrow(
+                            () -> new NotFoundException("The workshop id " + id +" was not found.")
+                    );
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public Iterable<WorkshopModel> findAllWorkshops() {
-        return repository.findAll();
+    public Iterable<WorkshopModel> findAllWorkshops() throws ExpiredJwtException {
+        try{
+            return repository.findAll();
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public WorkshopModel findWorkshopById(Long id) {
-        return findOrThrow(id);
+    public WorkshopModel findWorkshopById(Long id) throws ExpiredJwtException {
+        try{
+            if(!repository.existsById(id)) throw new NotFoundException("The workshop id " + id +" was not found.");
+            return findOrThrow(id);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public void deleteWorkshopById(Long id) {
-        repository.deleteById(id);
+    public void deleteWorkshopById(Long id) throws ExpiredJwtException {
+        try{
+            if(!repository.existsById(id)) throw new NotFoundException("The workshop id " + id +" was not found.");
+            repository.deleteById(id);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public WorkshopModel createWorkshop(WorkshopModel workshop) {
-        return repository.save(workshop);
+    public WorkshopModel createWorkshop(WorkshopModel workshop) throws ExpiredJwtException {
+        try{
+            return repository.save(workshop);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 
-    public void updateWorkshop(Long id, WorkshopModel workshop) {
-        findOrThrow(id);
-        repository.save(workshop);
+    public void updateWorkshop(Long id, WorkshopModel workshop) throws ExpiredJwtException {
+        try{
+            if(!repository.existsById(id)) throw new NotFoundException("The workshop id " + id +" was not found.");
+            findOrThrow(id);
+            repository.save(workshop);
+        } catch (Exception e) {
+            throw new ExpiredJwtException("You are not authorized to access this resource.");
+        }
     }
 }
